requirements.txt
fastapi
uvicorn
matplotlib
numpy
scipy
sympy
seaborn
imageio
pillow
pydantic
pytest
httpx
